Sketch moving oneshot executor
