+++ core/lib/vm_executor/src/lib.rs
@@ -6,4 +6,6 @@
 pub use zksync_multivm::interface::executor as interface;
 
 pub mod batch;
+pub mod oneshot;
+mod shared;
 pub mod storage;
