+++ core/lib/vm_interface/src/executor.rs
@@ -6,8 +6,9 @@ use async_trait::async_trait;
 use zksync_types::Transaction;
 
 use crate::{
-    storage::StorageView, BatchTransactionExecutionResult, FinishedL1Batch, L1BatchEnv, L2BlockEnv,
-    SystemEnv,
+    storage::{ReadStorage, StorageView},
+    BatchTransactionExecutionResult, BytecodeCompressionError, FinishedL1Batch, L1BatchEnv,
+    L2BlockEnv, OneshotEnv, SystemEnv, TxExecutionArgs, VmExecutionResultAndLogs,
 };
 
 /// Factory of [`BatchExecutor`]s.
@@ -41,4 +42,29 @@ pub trait BatchExecutor<S>: 'static + Send + fmt::Debug {
 
     /// Finished the current L1 batch.
     async fn finish_batch(self: Box<Self>) -> anyhow::Result<(FinishedL1Batch, StorageView<S>)>;
+}
+
+/// VM executor capable of executing isolated transactions / calls (as opposed to [batch execution](BatchExecutor)).
+#[async_trait]
+pub trait OneshotExecutor<S: ReadStorage> {
+    type Tracers: Default; // FIXME: revise
+
+    async fn inspect_transaction(
+        &self,
+        storage: S,
+        env: OneshotEnv,
+        args: TxExecutionArgs,
+        tracers: Self::Tracers,
+    ) -> anyhow::Result<VmExecutionResultAndLogs>;
+
+    async fn inspect_transaction_with_bytecode_compression(
+        &self,
+        storage: S,
+        env: OneshotEnv,
+        args: TxExecutionArgs,
+        tracers: Self::Tracers,
+    ) -> anyhow::Result<(
+        Result<(), BytecodeCompressionError>,
+        VmExecutionResultAndLogs,
+    )>;
 }
